refactor: product opiton과 quantity 디렉토리 위치 product 공용으로 이동
